Replace stage 3 wget with busybox's wget
Stage 3 wget has issues with linking to some libraries (some libs in
lib64 are not copied to the initramfs which causes "relocation errors"
in wget).

So we have replaced this wget with the one from busybox. Unfortunately
this means we lose https support, since busybox's version does not
support it (as such, all https links have been changed to http).